1-API: Move API response DTOs into `Shared` folder
